product_api/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
python-decouple==3.8
asgiref==3.8.1
django-filter==24.3
mysqlclient==2.2.6
pillow==11.0.0
sqlparse==0.5.3
tzdata==2024.2
pycodestyle==2.12.1
pyflakes==3.2.0